[OPENMEETINGS-2014] user picture is displayed as expected
